The special equals method and the Interval class to support it were a red herring and a dead end -- the problem was just the IntervalTree's own lookup method. So I reverted everything to use Range.
I haven't killed Interval yet because a few helper methods were defined in it, but I'll extract them and transform its calls on functions where they are used.

The code now actually passes some sample tests, which is a great improvement, and an possible indication I'm on the right track.